Unauthenticated user chartering snackbar warning
Fixed issue of authenticated user being able to access the chartering page, and added functionality to redirect a logged out user to the landing page if they attempt to access the chartering page while logged out. Also added a snackbar popup to warn the user of this issue. Removed React's strictmode so that the snackbar popup doesn't appear twice.