Updated code to be more performant
Instead of checking for input for EVERY form field, changed it to one call in Create.tsx to check if there is any unsaved data. Reload prompt functionality fully tested and working. All that's left to do is to add functionality to navigation or route change causing the browser to prompt the user.