Added 404 page functionality
Made a simple 404 page that shows whenever the user enters an invalid URL, with a button to return to the homepage. For some reason, there is a slight delay before the page shows up, perhaps the site searching the database for any organizations with the entered name. Hopefully the 404 page will look better after a UI overhaul.